Move zoomed image to components
