+++ content/webapp/components/CaptionedImage/CaptionedImage.tsx
@@ -4,8 +4,7 @@ import styled from 'styled-components';
 import { CaptionedImage as CaptionedImageType } from '@weco/common/model/captioned-image';
 import ImageWithTasl from '../ImageWithTasl/ImageWithTasl';
 import HeightRestrictedPrismicImage from '@weco/common/views/components/HeightRestrictedPrismicImage/HeightRestrictedPrismicImage';
-import Icon from '@weco/common/views/components/Icon/Icon';
-import { expand, cross } from '@weco/common/icons';
+import { ZoomedImage, ZoomedImageButton } from '../ZoomedImage/ZoomedImage';
 
 type CaptionedImageProps = {
   isBody?: boolean;
@@ -59,45 +58,6 @@ const ImageContainerInner = styled.div<ImageContainerInnerProps>`
   }
 `;
 
-const ZoomButton = styled.button`
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  z-index: 1;
-  color: ${props => props.theme.color('white')};
-  background: ${props => props.theme.color('black')};
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 0;
-  padding: 0;
-  margin: 0;
-`;
-
-const StyledDialog = styled.dialog`
-  border: 0;
-  padding: 0;
-  max-width: 100%;
-  max-height: 100%;
-  background: ${props => props.theme.color('black')};
-
-  img {
-    width: 100vw;
-    height: 100vh;
-    display: block;
-    object-fit: contain;
-  }
-
-  button {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-  }
-`;
-
 type UiCaptionedImageProps = CaptionedImageType & {
   isBody?: boolean;
   preCaptionNode?: ReactNode;
@@ -110,17 +70,9 @@ const CaptionedImage: FC<UiCaptionedImageProps> = ({
   isBody,
   hasRoundedCorners,
 }) => {
-  const dialogRef: RefObject<HTMLDialogElement> = useRef(null);
-  const [didOpenModal, setDidOpenModal] = useState(false);
+  const zoomRef: RefObject<HTMLDialogElement> = useRef(null);
+  const [isZoom, setIsZoom] = useState(false);
 
-  function openDialog() {
-    dialogRef?.current?.showModal();
-    setDidOpenModal(true);
-  }
-
-  function closeDialog() {
-    dialogRef?.current?.close();
-  }
   // Note: the default quality here was originally 45, but this caused images to
   // appear very fuzzy on stories.
   //
@@ -136,25 +88,14 @@ const CaptionedImage: FC<UiCaptionedImageProps> = ({
         aspectRatio={image.width / image.height}
         hasRoundedCorners={hasRoundedCorners}
       >
-        <ZoomButton onClick={openDialog}>
-          <Icon icon={expand} color="white" />
-        </ZoomButton>
+        <ZoomedImageButton setIsZoom={setIsZoom} zoomRef={zoomRef} />
         <ImageWithTasl
           Image={<HeightRestrictedPrismicImage image={image} quality="high" />}
           tasl={image.tasl}
         />
         <Caption caption={caption} preCaptionNode={preCaptionNode} />
       </ImageContainerInner>
-      <StyledDialog ref={dialogRef}>
-        {didOpenModal && (
-          <>
-            <ZoomButton onClick={closeDialog}>
-              <Icon icon={cross} color="white" />
-            </ZoomButton>
-            <img src={image.contentUrl} alt={image.alt || ''} />
-          </>
-        )}
-      </StyledDialog>
+      <ZoomedImage image={image} zoomRef={zoomRef} isZoom={isZoom} />
     </CaptionedImageFigure>
   );
 };
